Second

+++ src/main/java/system/university/pds/model/Order.java
@@ -1,15 +1,11 @@
 package system.university.pds.model;
 
-import lombok.Getter;
-
-import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public class Order {
-    private Map<BigInteger, Course> courses;
+    private Map<Integer, Course> courses;
     private final User student;
 
     public Order(User student) {
@@ -21,7 +17,7 @@ public class Order {
         return (List<Course>) courses.values();
     }
 
-    public void addCourse(BigInteger courseId, Course course){
+    public void addCourse(int courseId, Course course){
         courses.put(courseId, course);
     }
 }
